Move Conditions to conditions.py
